backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
requests==2.31.0
pydantic==2.7.4
pydantic-settings==2.2.1
chromadb==0.5.3
pandas==2.2.2
numpy==1.26.4
scikit-learn==1.4.2
tensorflow==2.16.1
aiohttp==3.13.3
loguru==0.7.2
python-multipart==0.0.9
python-dotenv==1.0.1
joblib==1.3.2
pytest==8.2.2
pytest-asyncio==0.22.0
openai==1.35.3
langchain-community==0.0.38
docker==6.1.3
boto3==1.34.0
google-generativeai==0.7.1
frozenlist==1.4.1
